add --base-dir option
This is equivalent to the -basedir flag in the original game. It
indicates to the engine the parent of the id1/ directory. If the
id1/ directory is not found, Richter now helpfully logs an error
explaining the problem and how to solve it. This is an improvement
over the previous behavior, shared by the original game, which
simply presents the cryptic message "gfx/palette.lmp not found".