[wip] update for 2024
This is mostly a maintenance update, with the goal of getting the client
to build and run again. The winit and wgpu dependencies have been
updated to their most recent versions, as the old versions wouldn't play
nice with current GPU drivers.

The game still doesn't run -- it looks like either shaderc or Naga is
failing to handle offset specifiers in push constant blocks. Hopefully
that's fixable without rewriting the shaders in WGSL...